Merge 9.0 PostgreSQL release into PGXC. Resolve conflicts thrown by git
and fix some issues raised during compilation. We still don't compile fine
at this point, but we should have resolved many conflicts to make further
progress.

Some of the changes in the regression tests are merged to reflect whats
there in 9.0 release. Those are easy to fix later when we run regressions

Conflicts:
	contrib/Makefile
	contrib/pgbench/pgbench.c
	src/Makefile
	src/backend/Makefile
	src/backend/access/transam/varsup.c
	src/backend/catalog/Makefile
	src/backend/catalog/dependency.c
	src/backend/catalog/genbki.sh
	src/backend/commands/dbcommands.c
	src/backend/commands/explain.c
	src/backend/commands/vacuum.c
	src/backend/executor/execMain.c
	src/backend/executor/execProcnode.c
	src/backend/executor/execTuples.c
	src/backend/parser/analyze.c
	src/backend/parser/gram.y
	src/backend/parser/parse_utilcmd.c
	src/backend/postmaster/postmaster.c
	src/backend/rewrite/rewriteHandler.c
	src/backend/storage/ipc/procarray.c
	src/backend/storage/lmgr/proc.c
	src/backend/tcop/postgres.c
	src/backend/tcop/utility.c
	src/backend/utils/cache/relcache.c
	src/backend/utils/init/postinit.c
	src/backend/utils/misc/guc.c
	src/bin/pg_ctl/pg_ctl.c
	src/include/Makefile
	src/include/access/twophase.h
	src/include/bootstrap/bootstrap.h
	src/include/catalog/catversion.h
	src/include/catalog/dependency.h
	src/include/catalog/indexing.h
	src/include/catalog/pg_proc.h
	src/include/nodes/nodes.h
	src/include/storage/lwlock.h
	src/include/storage/proc.h
	src/include/storage/procarray.h
	src/include/utils/lsyscache.h
	src/test/regress/expected/delete.out
	src/test/regress/expected/float4.out
	src/test/regress/expected/float8.out
	src/test/regress/expected/geometry.out
	src/test/regress/expected/join.out
	src/test/regress/expected/point.out
	src/test/regress/expected/rowtypes.out
	src/test/regress/expected/timestamp.out
	src/test/regress/expected/timestamptz.out
	src/test/regress/expected/tsearch.out
	src/test/regress/sql/numeric.sql
	src/test/regress/sql/point.sql

+++ src/backend/nodes/read.c
@@ -4,12 +4,12 @@
  *	  routines to convert a string (legal ascii representation of node) back
  *	  to nodes
  *
- * Portions Copyright (c) 1996-2009, PostgreSQL Global Development Group
+ * Portions Copyright (c) 1996-2010, PostgreSQL Global Development Group
  * Portions Copyright (c) 1994, Regents of the University of California
  *
  *
  * IDENTIFICATION
- *	  $PostgreSQL: pgsql/src/backend/nodes/read.c,v 1.52 2009/01/01 17:23:43 momjian Exp $
+ *	  $PostgreSQL: pgsql/src/backend/nodes/read.c,v 1.54 2010/01/18 22:19:34 petere Exp $
  *
  * HISTORY
  *	  AUTHOR			DATE			MAJOR EVENT
@@ -289,7 +289,7 @@ nodeRead(char *token, int tok_len)
 
 	type = nodeTokenType(token, tok_len);
 
-	switch (type)
+	switch ((int) type)
 	{
 		case LEFT_BRACE:
 			result = parseNodeString();
